Renamed module core to qtcore.
`examples/kefia` and `examples/rqbg` are currently broken. All other
tests work fine.
